twitch_bot/requirements.txt
twitchio
python-dotenv
requests
twitchAPI